% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Blue.comb}
\alias{Blue.comb}
\title{Blueprint+Encode reference dataset for human}
\format{
A list with the following items:
\describe{
\item{data}{Gene expression data matrix}
\item{types}{vector of annotation per column in the matrix}
\item{main_types}{vector of broad annotations per column in the matrix}
\item{name}{name if the reference set}
\item{sd.thres}{threshold of the standard deviation for genes to use in SingleR}
\item{de.genes.main}{list of lists of differentially expressed genes between every two cell types in main_types}
\item{de.genes}{list of lists of differentially expressed genes between every two cell types in types}
}
}
\usage{
Blue.comb
}
\description{
Blueprint+Encode reference dataset for human
}
\keyword{datasets}
