% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STCNV.R
\name{CNV_heatmap}
\alias{CNV_heatmap}
\title{Create a CNV Heatmap}
\usage{
CNV_heatmap(adata, output_dir, file_name = "D8_rm_leiden3_cnv.pdf")
}
\arguments{
\item{adata}{An input data object containing information for CNV analysis.}

\item{output_dir}{The directory where the output heatmap file will be saved.}

\item{file_name}{The name of the output heatmap file (default is "D8_rm_leiden3_cnv.pdf").}
}
\value{
This function doesn't return a value but saves the CNV heatmap as a PDF file.
}
\description{
This function generates a CNV (Copy Number Variation) heatmap using input data.
}
