% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneOrder}
\alias{geneOrder}
\title{Gene order}
\format{
A dataframe with the following items:
\describe{
\item{seqnames}{chromosome names}
\item{start}{gene start}
\item{end}{gene end}
}
}
\usage{
geneOrder
}
\description{
Gene order
}
\keyword{datasets}
