% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STpheno.R
\name{mmr_loss_score}
\alias{mmr_loss_score}
\title{Calculate Score, P-Value, and Call for a Single Gene}
\usage{
mmr_loss_score(x, m, sd, p.thresh = 0.01)
}
\arguments{
\item{x}{The expression value of the gene.}

\item{m}{The mean expression value of the gene.}

\item{sd}{The standard deviation of the gene's expression values.}

\item{p.thresh}{The p-value threshold for the call (default: 0.01).}
}
\value{
A list containing the following components:
\describe{
\item{score}{The calculated score for the gene, representing its deviation from the mean, scaled by the standard deviation implied by the interquartile range (IQR).}
\item{pval}{The p-value associated with the gene's score.}
\item{call}{A binary call (0 or 1) indicating whether the gene's p-value is below the specified threshold.}
}
}
\description{
This function calculates a score, p-value, and call for a single gene based on its expression value.
}
