% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmr_matrix}
\alias{mmr_matrix}
\title{MMR gene expression matrix}
\format{
A matrix with the following items:
\describe{
\item{colnames}{sample id}
\item{rownames}{MMR genes}
\item{value}{mean expression in malignant cells}
}
}
\usage{
mmr_matrix
}
\description{
MMR gene expression matrix
}
\keyword{datasets}
