% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STanno.R
\name{STanno}
\alias{STanno}
\title{Perform celltype annotation}
\usage{
STanno(seurat_obj, cutoff = 0.1)
}
\arguments{
\item{seurat_obj}{A Seurat object containing spatial transcriptome data.}

\item{cutoff}{A cutoff value of AUC Score to determine endothelial cells (default: 0.1).}
}
\value{
A modified Seurat object with additional cell type annotations.
}
\description{
This function performs celltype annotation, including vascular signature scoring, cell type assignment, and spatial visualization.
}
