% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STCNV.R
\name{STCNV}
\alias{STCNV}
\title{Perform spatial transcriptome Copy Number Variation (CNV) Analysis}
\usage{
STCNV(
  seurat_obj,
  test_adata,
  output_dir,
  gene_order_file,
  metadata,
  patient_id,
  ref_group_names,
  cutoff = 0.01,
  cluster_by_groups = T,
  denoise = T,
  HMM = F,
  num_threads = 15,
  tumor_subcluster_partition_method = "qnorm",
  analysis_mode = "samples",
  k_nn = 30,
  leiden_resolution = 1,
  tumor_subcluster_pval = 0.05,
  iter_max = 100,
  file_name = "D8_rm_leiden3_cnv.pdf"
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing spatial transcriptome data.}

\item{test_adata}{A Seurat object containing test spatial transcriptome data.}

\item{output_dir}{The directory where the output files and results will be saved.}

\item{gene_order_file}{A .Rdata file containing gene order information.}

\item{metadata}{A vector specifying group variables to identify test and ref.}

\item{patient_id}{A unique identifier for the patient or sample being analyzed.}

\item{ref_group_names}{A character vector of reference group names for CNV estimation.}

\item{cutoff}{Cut-off for the min average read counts per gene among reference cells (default: 0.01).}

\item{cluster_by_groups}{If observations are defined according to groups (ie. patients), each group of cells will be clustered separately. (default=FALSE, instead will use k_obs_groups setting) (default: TRUE).}

\item{denoise}{A logical indicating whether to perform data denoising (default: TRUE).}

\item{HMM}{A logical indicating whether to use Hidden Markov Models (HMM) for CNV inference (default: FALSE).}

\item{num_threads}{The number of CPU threads to use for parallel processing (default: 15).}

\item{tumor_subcluster_partition_method}{The method used for partitioning tumor subclusters (default: 'qnorm').}

\item{analysis_mode}{The mode of analysis ('samples' or 'subclusters' or 'cells'; default: 'samples').}

\item{k_nn}{The number of nearest neighbors for clustering (default: 30).}

\item{leiden_resolution}{The resolution parameter for Leiden clustering (default: 1).}

\item{tumor_subcluster_pval}{The p-value threshold for tumor subcluster identification (default: 0.05).}

\item{iter_max}{The maximum number of iterations for k-means clustering (default: 100).}

\item{file_name}{file name to save the heatmap plot.}
}
\value{
A Seurat object with additional metadata columns (cnv_score, malignant/non-malignant).
\itemize{
\item CNV_Score.txt: CNV scores in spatial.
\item kmeans_df_s.txt: Results of k-means clustering.
\item spatial_CNV_score.png: Spatial plot of CNV scores.
\item spatial_kmeans.png: Spatial plot of k-means clustering results.
\item vln_cnv_score_kmeans.png: Violin plot of CNV scores by k-means clusters.
\item spatial_malignant.pdf: Spatial plot of malignant cell identification.
\item vln_cnv_score_malig.png: Violin plot of CNV scores between malignant/non-malignant.
}
}
\description{
This function conducts a comprehensive analysis of spatial transcriptome copy number variations (CNV) using Seurat and InferCNV.
}
