% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STpheno.R
\name{STpheno}
\alias{STpheno}
\title{Perform MMR Loss Score Analysis for Multiple Seurat Objects}
\usage{
STpheno(seurat_obj = NULL, sample_id, phenotype_id)
}
\arguments{
\item{seurat_obj}{A list of Seurat objects, each representing spatial transcriptome data for different samples.}

\item{sample_id}{A character vector of sample IDs corresponding to the Seurat objects.}

\item{phenotype_id}{A character vector of phenotype labels for each sample.}
}
\value{
This function generates various plots and does not have a formal return value.
}
\description{
This function performs MMR (Mismatch Repair) Loss Score analysis for multiple Seurat objects, specifically related to colorectal cancer (CRC) phenotypes.
}
